vision_service/requirements.txt
fastapi==0.115.5
uvicorn==0.30.6
ultralytics==8.2.82
pillow==10.4.0
